Fixed the bug with the mesher, which didn't produce proper mesh elements when more than 2 boundary conditions were added

+++ core/computationalCore/problem/include/problem/boundaryPatch.h
@@ -19,5 +19,5 @@ struct BoundaryPatch {
 	BoundaryConditionType type;
 
 	CudaArray<uint32_t> faceIDs;
-	CudaArray<uint32_t> values;
+	CudaArray<double> values;
 };
